+++ src/LighthouseServiceProvider.php
@@ -14,6 +14,8 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Support\ServiceProvider;
 use Nuwave\Lighthouse\Console\CacheCommand;
 use Nuwave\Lighthouse\Console\ClearCacheCommand;
+use Nuwave\Lighthouse\Console\ClearQueryCacheCommand;
+use Nuwave\Lighthouse\Console\ClearSchemaCacheCommand;
 use Nuwave\Lighthouse\Console\DirectiveCommand;
 use Nuwave\Lighthouse\Console\FieldCommand;
 use Nuwave\Lighthouse\Console\IdeHelperCommand;
@@ -56,6 +58,8 @@ class LighthouseServiceProvider extends ServiceProvider
     protected const COMMANDS = [
         CacheCommand::class,
         ClearCacheCommand::class,
+        ClearQueryCacheCommand::class,
+        ClearSchemaCacheCommand::class,
         DirectiveCommand::class,
         FieldCommand::class,
         IdeHelperCommand::class,