+++ CHANGELOG.md
@@ -9,6 +9,14 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Add a file based query cache to leverage OPcache for improved performance https://github.com/nuwave/lighthouse/pull/2713
+
+### Deprecated
+
+- Deprecate command `lighthouse:clear-cache` in favor of `lighthouse:clear-schema-cache` https://github.com/nuwave/lighthouse/pull/2713
+
 ## v6.62.3
 
 ### Fixed